fix(provider): restore the quota test the split overwrote
discovery_quota_test.go ALREADY EXISTED, holding
TestFetchQuotaJSON_InvalidBaseURL: it drives the public DeepSeek balance
flow with a control character in the provider base URL and asserts the
request construction fails rather than panicking. My splitter wrote the
quota bucket of discovery_test.go to that same path and destroyed it, so
the previous commit's "pure moves, no test body changes" was untrue for
this one file.

The original file is restored verbatim, and the bucket that clobbered it
(the provider cache round-trip and expiry, plus the quota circuit
breaker and its retry policy) moves to discovery_cache_test.go, which is
also the better name for what it holds.

Reported by Greptile on #793, with a reproduction that ran the deleted
case against the public path.
